Capture refactoring
