+++ src/Fluent.Calculations.Primitives/Expressions/Capture/CapturedEvaulation.cs
@@ -1,10 +1,10 @@
-﻿namespace Fluent.Calculations.Primitives.Expressions;
+﻿namespace Fluent.Calculations.Primitives.Expressions.Capture;
 
 internal class CapturedEvaulation
 {
     public string Name { get; }
 
-    public CapturedEvaulation(string name) => this.Name = name;
+    public CapturedEvaulation(string name) => Name = name;
 
     internal static bool OfType(object obj) => obj is CapturedEvaulation;
 }
